% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sxpdb.R
\name{add_origin}
\alias{add_origin}
\title{Add origin to an already recorded value}
\usage{
add_origin(db, hash, package, func, argument)
}
\arguments{
\item{db}{database, sxpdb object}

\item{hash}{hash of an already recorded value}

\item{package}{character vector for the package name}

\item{func}{character vector for the function name}

\item{argument}{character vector for the argument name. \code{""} or \code{NA} mean that \code{val} is a return value.}
}
\value{
\code{NULL}
}
\description{
\code{add_origin} adds package, function and argument names for a value already recorded in the database.
The value is represented here by its hash. You will not probably use it and rather use \code{\link[=add_val_origin]{add_val_origin()}}
directly, as \code{add_origin} will only save you the hash computation. If performance is an issue, you
should not use the R API anyway and directly hook into the C one.
}
\seealso{
\code{\link[=add_val_origin]{add_val_origin()}}, \code{\link[=build_indexes]{build_indexes()}}
}
