% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sxpdb.R
\name{add_val}
\alias{add_val}
\title{Add a value in the dabatase}
\usage{
add_val(db, val)
}
\arguments{
\item{db}{database, sxpdb object}

\item{val}{any R value. Currently, environments and closures will be silently ignored}
}
\value{
integer index of the value if it has been added (now or before), or \code{NULL} if it has been ignored
}
\description{
\code{add_val} adds an R value in the database. It does not add origins or call ids.
This should be used in conjunction with \code{\link[=add_origin]{add_origin()}}.
You should most likely never use it and rather directly use \code{\link[=add_val_origin]{add_val_origin()}}.
}
\seealso{
\code{\link[=add_val_origin]{add_val_origin()}}, \code{\link[=add_origin]{add_origin()}}, \code{\link[=build_indexes]{build_indexes()}}
}
