% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sxpdb.R
\name{add_val_origin}
\alias{add_val_origin}
\title{Add a value with origin and call id}
\usage{
add_val_origin(db, val, package, func, argument, call_id = 0)
}
\arguments{
\item{db}{database, sxpdb object}

\item{val}{any R value. Currently, environments and closures will be silently ignored}

\item{package}{character vector for the package name}

\item{func}{character vector for the function name}

\item{argument}{character vector for the argument name. \code{""} or \code{NA} mean that \code{val} is a return value.}

\item{call_id}{integer unique id of the call from which the value comes from. Defaults to \code{0}}
}
\value{
integer index of the value if it has been added (now or before), or \code{NULL} if it has been ignored
}
\description{
\code{add_val_origin} adds an R value in the database, along with its origin
(package, function, argument names), and its call id.
}
\seealso{
\code{\link[=build_indexes]{build_indexes()}}
}
