% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sxpdb.R
\name{build_indexes}
\alias{build_indexes}
\title{Build search indexes.}
\usage{
build_indexes(db)
}
\arguments{
\item{db}{database, sxpdb object}
}
\value{
\code{NULL}
}
\description{
\code{build_indexes} explicitly builds search indexes which are used by function with a \code{query} argument
(when it is not \code{NULL}). You need to use it if you add new values into the database (including merging into it).
However, \code{\link[=merge_all_dbs]{merge_all_dbs()}} will automatically build the search indexes.
}
