% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sxpdb.R
\name{close_db}
\alias{close_db}
\title{Closes the database}
\usage{
close_db(db)
}
\arguments{
\item{db}{database, sxpdb object}
}
\value{
\code{NULL}
}
\description{
Closes a previously open database. If \code{db} was uncorrectly opened, or has been already closed,
the function will error.
}
\seealso{
\code{\link[=open_db]{open_db()}}
}
