% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sxpdb.R
\name{close_query}
\alias{close_query}
\title{Closes a query object.}
\usage{
close_query(query)
}
\arguments{
\item{query}{query object}
}
\value{
\code{NULL}
}
\description{
\code{close_query} closes a query object. A query object is implemented as an external pointer to a
a C++ object that he dynamically allocated. However, a GC hook for that external pointer is also
registered so except if you plan to crate a huge amount of query objects, you should be fine
with not calling \code{close_query} explicitly.
}
\seealso{
\code{\link[=query_from_value]{query_from_value()}}, \code{\link[=query_from_plan]{query_from_plan()}}, \code{\link[=relax_query]{relax_query()}}
}
