% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sxpdb.R
\name{filter_index_db}
\alias{filter_index_db}
\title{Filters values from the database according to some predicate}
\usage{
filter_index_db(db, fun, query = NULL)
}
\arguments{
\item{db}{database, sxpdb object}

\item{fun}{R function, the predicate, which should take one argument, the value, and return a boolean.
ǸA_logical_\verb{is considered as}TRUE`.}

\item{query}{query object, typically built from \code{\link[=query_from_plan]{query_from_plan()}} or \code{\link[=query_from_value]{query_from_value()}}.}
}
\value{
list of indices of the values matche2d by the query for which \code{fun} evaluated to \code{TRUE}
or \code{NA_logical_}.
}
\description{
Sometimes, the query language is not enough to select some values. You can use \code{filter_index_db}
to refine a query and only keep the values for which the predicate is true.
}
\seealso{
\code{\link[=map_db]{map_db()}}, \code{\link[=view_db]{view_db()}}
}
