% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sxpdb.R
\name{get_origins}
\alias{get_origins}
\title{Get the origins of a value}
\usage{
get_origins(db, hash)
}
\arguments{
\item{db}{database, sxpdb object}

\item{hash}{RAW, hash of the value to look for}
}
\value{
data frame with columns \code{pkg}, \code{fun} and \code{param} for  package
function and parameter names.
}
\description{
\code{get_origins} returns the origins (package, function, parameter names) of a function given the hash
of the value. It save a hash computation, compared to \code{\link[=get_origins_idx]{get_origins_idx()}}.
}
\seealso{
\code{\link[=get_origins_idx]{get_origins_idx()}}, \code{\link[=view_origins_db]{view_origins_db()}}
}
