% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sxpdb.R
\name{get_origins_idx}
\alias{get_origins_idx}
\title{Get the origins of a value}
\usage{
get_origins_idx(db, i)
}
\arguments{
\item{db}{database, sxpdb object}

\item{i}{integer, index of the value in the database}
}
\value{
data frame with columns \code{pkg}, \code{fun} and \code{param} for  package
function and parameter names.
}
\description{
\code{get_origins} returns the origins (package, function, parameter names) of a function given the index
of the value.
}
\seealso{
\code{\link[=get_origins]{get_origins()}}, \code{\link[=view_origins_db]{view_origins_db()}}
}
