% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sxpdb.R
\name{get_value_idx}
\alias{get_value_idx}
\title{Get the value in the db at a given index}
\usage{
get_value_idx(db, idx)
}
\arguments{
\item{db}{database, sxpdb object}

\item{idx}{integer, index in the database. Indexes start at 0 (not at 1!) so \code{idx >= 0} and
\code{idx < size_db(db)}}
}
\value{
an R value
}
\description{
\code{get_value_idx} returns the value in \code{db} at index \code{idx}. You can also use the notation \code{db[idx]}.
}
\seealso{
[\verb{[.sxpdb}()]
}
