% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sxpdb.R
\name{has_search_index}
\alias{has_search_index}
\title{Checks if the database has a search index}
\usage{
has_search_index(db)
}
\arguments{
\item{db}{database, sxpdb object}
}
\value{
boolean, \code{TRUE} if it has a search index
}
\description{
Requests to the database that pass a non-NULL query argument will
not work if the search index is not built.
}
\seealso{
\code{\link[=build_indexes]{build_indexes()}}
}
