% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sxpdb.R
\name{have_seen}
\alias{have_seen}
\title{Checks if a value is in the database}
\usage{
have_seen(db, val)
}
\arguments{
\item{db}{database, sxpdb object}

\item{val}{R value}
}
\value{
integer index of the value if the value is in the db, \code{NULL} otherwise
}
\description{
\code{have_seen} checks if a value is stored in the database and returns its index in that case.
}
