% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sxpdb.R
\name{is_query_empty}
\alias{is_query_empty}
\title{Checks if a query object is empty.}
\usage{
is_query_empty(query)
}
\arguments{
\item{query}{query object}
}
\value{
boolean, whether the query is empty or not
}
\description{
\code{is_query_empty} checks if the query refers to some values, or no values at all. You need to call it after at least
It is typically used when the function using the query returns an ambiguous result (e.g. \code{\link[=sample_val]{sample_val()}})
and you want to distinguish between \code{NULL} and an empty query that results also in \code{NULL}.
}
\seealso{
\code{\link[=query_from_value]{query_from_value()}}, \code{\link[=query_from_plan]{query_from_plan()}}, \code{\link[=relax_query]{relax_query()}}, \code{\link[=close_query]{close_query()}}
}
