% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sxpdb.R
\name{map_db}
\alias{map_db}
\title{Map a function on the values in the database}
\usage{
map_db(db, fun, query = NULL)
}
\arguments{
\item{db}{database, sxpdb object}

\item{fun}{R function, the function to apply to each value matching the query. It takes on argument,
the value and should return an R value.}

\item{query}{query object, typically built from \code{\link[=query_from_plan]{query_from_plan()}} or \code{\link[=query_from_value]{query_from_value()}}.}
}
\description{
\code{map_db} runs a function on each of the elements of the database matching a query
and return a list of the results.
}
\seealso{
\code{\link[=view_db]{view_db()}}, \code{\link[=filter_index_db]{filter_index_db()}}, \code{\link[=view_meta_db]{view_meta_db()}}, \code{\link[=view_origins_db]{view_origins_db()}}
}
