% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sxpdb.R
\name{merge_all_dbs}
\alias{merge_all_dbs}
\title{Merges several dbs into a new large one.}
\usage{
merge_all_dbs(db_paths, output_path, parallel = TRUE)
}
\arguments{
\item{db_paths}{character vector of paths to the databases to be merged together}

\item{output_path}{character vector of the path of the resulting database}

\item{parallel}{boolean, whether the merge should be performed in parallel or not.}
}
\value{
data frame with information about the merging process. The data frame has the following columns:
\itemize{
\item \code{path}: path of the merged database
\item \code{db_size_before}:  size of the resulting database before merging the db at \code{path}
\item \code{added_values}: number of new unique values added after merging the db at \code{path}
\item \code{small_db_size}: number of values in the db at \code{path}
\item \code{small_db_bytes}: size in bytes of the db at \code{path}
\item \code{duration}: duration in seconds of merging the db at \code{path} into the resulting db
\item \code{error}: whether there was an error when merging the db at \code{path}.
}
}
\description{
\code{merge_all_dbs} performs an efficient merge of several databases, to keep only
unique values.
}
\seealso{
\code{\link[=merge_into]{merge_into()}}
}
