% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sxpdb.R
\name{merge_db}
\alias{merge_db}
\title{Merge a db into another one.}
\usage{
merge_db(db1, db2)
}
\arguments{
\item{db1}{database, sxpdb object}

\item{db2}{database, sxpdb object}
}
\value{
\code{NULL} in case of error, number of values in the \emph{target} database after merging otherwise
}
\description{
Deprecated. Rather use \code{\link[=merge_into]{merge_into()}}
}
