% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sxpdb.R
\name{merge_into}
\alias{merge_into}
\title{Merge a source db into a target db}
\usage{
merge_into(target, source)
}
\arguments{
\item{target}{database, sxpdb object}

\item{source}{database, sxpdb object}
}
\value{
\code{NULL}in case of an error, a mapping from old indexes of all the values in the \emph{source}
database to the new indexes for them in the \emph{target} database.
}
\description{
\code{merge_into} merges db \emph{source} into db \emph{target}. Db \emph{target} must be open in write mode and
db \emph{source} must be opened in merge mode.
}
\seealso{
\code{\link[=merge_all_dbs]{merge_all_dbs()}}, \code{\link[=build_indexes]{build_indexes()}}
}
