% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sxpdb.R
\name{nb_values_db}
\alias{nb_values_db}
\title{Computes the number of values matching a given query}
\usage{
nb_values_db(db, query = NULL)
}
\arguments{
\item{db}{database, sxpdb object}

\item{query}{query object or \code{NULL}. If \code{NULL}, samples from the whole database}
}
\value{
integer, number of values matching the query.
}
\description{
\code{nb_values_db} returns the number of values matching a given query. This is very cheap to compute
as it will just intersect or union some pre-computed arrays, so prefer to using \code{\link[=view_meta_db]{view_meta_db()}}
for simple quantitative questions. \code{nb_values_db} called with a \code{NULL} query is equivalent to calling
\code{size_db}.
}
\seealso{
\code{\link[=size_db]{size_db()}}
}
