% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sxpdb.R
\name{open_db}
\alias{open_db}
\title{Open a database}
\usage{
open_db(db = "db", mode = FALSE, quiet = TRUE)
}
\arguments{
\item{db}{character vector of the name and path where to create the db. Default to \code{"db"}}

\item{mode}{\code{TRUE} if in write mode, \code{FALSE} if in read mode (default), \code{"merge"} if in merge mode}

\item{quiet}{boolean, whether to print messages or not}
}
\value{
\code{NULL} on error, a external pointer with class tag "sxpdb" and class "sxpdb" to the database on success
}
\description{
\code{open_db} opens a database for reading, writing, or merging. Reading does not load as much in memory.
For instance, it will directly read values from the disk. When merging database \emph{source} into database
\emph{target}, \emph{source} must be open in merge mode, and \emph{target}, in write mode. Writing mode is required
when adding any new values to the database or building or updating search indexes.
}
\seealso{
\code{\link[=close_db]{close_db()}}
}
