% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sxpdb.R
\name{path_db}
\alias{path_db}
\title{Path to the database directory}
\usage{
path_db(db)
}
\arguments{
\item{db}{database, sxpdb object}
}
\value{
character path to the database
}
\description{
The database is stored as a directory of various configuration files and tables. \code{path_db}returns
the path to this directory.
}
