% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sxpdb.R
\name{query_from_plan}
\alias{query_from_plan}
\title{Creates a query from a plan.}
\usage{
query_from_plan(plan)
}
\arguments{
\item{plan}{named list describing the query. In the absence of a parameter name, the associated metadata
will be set to unspecified. The plan parameters are:
\itemize{
\item type: any value of the desired type
\item vector: boolean
\item length: integer
\item class name: character vector of class names, or boolean (to just specify that you want classes, but without specific class names)
\item na: boolean
\item ndims: integer
\item attributes: boolean
}}
}
\value{
query object
}
\description{
\code{query_from_plan} creates a query from a plan describing the query.  You can relax later on some
parameters of the query with \code{\link[=relax_query]{relax_query()}}.
The query can then be used with any function taking a \code{query} argument.
}
\seealso{
\code{\link[=query_from_value]{query_from_value()}}, \code{\link[=relax_query]{relax_query()}}, \code{\link[=close_query]{close_query()}}, \code{\link[=view_db]{view_db()}}, \code{\link[=map_db]{map_db()}}
}
