% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sxpdb.R
\name{query_from_value}
\alias{query_from_value}
\title{Creates a query from an example value.}
\usage{
query_from_value(value)
}
\arguments{
\item{value}{any R value}
}
\value{
query object
}
\description{
\code{query_from_value} creates a query from an example R value. The query will mimic the value
according to the following metadata: type, vector or not, length, class name, \code{NA} inside or not,
number of dimensions, attributes or not. You can then relax on some of those metadata to state
that you do not them want to be determined according to the example value, with \code{\link[=relax_query]{relax_query()}}.
The query can then be used with any function taking a \code{query} argument.
}
\seealso{
\code{\link[=relax_query]{relax_query()}}, \code{\link[=query_from_plan]{query_from_plan()}}, \code{\link[=close_query]{close_query()}}, \code{\link[=view_db]{view_db()}}, \code{\link[=map_db]{map_db()}}
}
