% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sxpdb.R
\name{relax_query}
\alias{relax_query}
\title{Relaxes on some parameters of a query.}
\usage{
relax_query(query, relax)
}
\arguments{
\item{query}{query object}

\item{relax}{character vector none or several of "na", "length", "attributes", "type", "vector", "ndims", "class". You can
also give "keep_type" or "keep_class" to relax on all constraints, except the type, or except the class names.}
}
\value{
boolean, \code{TRUE} if the query changed
}
\description{
\code{relax_query} makes a specified parameter from a query, i.e., with a given target value, into
a non-specified parameter. For instance, if the query specifies that it wants values with length
34, relaxing on the length parameter will allow values with any lengths,
not only values with lengths 34.
}
\seealso{
\code{\link[=query_from_value]{query_from_value()}}, \code{\link[=query_from_plan]{query_from_plan()}}, \code{\link[=close_query]{close_query()}}, \code{\link[=is_query_empty]{is_query_empty()}}
}
