% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sxpdb.R
\name{sample_index}
\alias{sample_index}
\title{Sample randomly a value from the database}
\usage{
sample_index(db, query = NULL)
}
\arguments{
\item{db}{database, sxpdb object}

\item{query}{query object or \code{NULL}. If \code{NULL}, samples from the whole database}
}
\value{
integer, an index to a value in the database. You can then access the value
with \code{\link[=get_value_idx]{get_value_idx()}}. \code{NULL} if the query is empty.
}
\description{
\code{sample_index} samples a value from the database and returns an index to the value.
The sampling uses an uniform distribution.
The value is picked along the whole database or according to a query built from \code{\link[=query_from_plan]{query_from_plan()}}
or \code{\link[=query_from_value]{query_from_value()}}.
}
\seealso{
\code{\link[=sample_val]{sample_val()}}
}
