% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sxpdb.R
\name{sample_similar}
\alias{sample_similar}
\title{Sample randomly a value similar to a given from the database}
\usage{
sample_similar(db, val, relax = "")
}
\arguments{
\item{db}{database, sxpdb object}

\item{val}{any R value}

\item{relax}{character vector, as in \code{\link[=relax_query]{relax_query()}}}
}
\value{
an R value. If the query is empty, it will return \code{NULL}. It is ambiguous with sampling
a value that  happens to be \code{NULL} so you should rather use \code{\link[=sample_index]{sample_index()}} with \code{\link[=query_from_value]{query_from_value()}}.
}
\description{
\code{sample_similar} samples a value from the database, which is similar to \code{val} along
metadata parameters, relaxed according to \code{relax}.
The sampling uses an uniform distribution.
}
\details{
You should rather use a combination of \code{\link[=query_from_value]{query_from_value()}} and \code{\link[=sample_index]{sample_index()}} for more control.
}
\seealso{
\code{\link[=sample_index]{sample_index()}}, \code{\link[=query_from_value]{query_from_value()}}
}
