% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sxpdb.R
\name{sample_val}
\alias{sample_val}
\title{Sample randomly a value from the database}
\usage{
sample_val(db, query = NULL)
}
\arguments{
\item{db}{database, sxpdb object}

\item{query}{query object or \code{NULL}. If \code{NULL}, samples from the whole database}
}
\value{
an R value. If the query is empty, it will return \code{NULL}. It is ambiguous with sampling
a value that  happens to be \code{NULL} so you should rather use \code{\link[=sample_index]{sample_index()}}.
}
\description{
\code{sample_val} samples a value from the database and returns the value.
The sampling uses an uniform distribution.
The value is picked along the whole database or according to a query built from \code{\link[=query_from_plan]{query_from_plan()}}
or \code{\link[=query_from_value]{query_from_value()}}.
}
\seealso{
\code{\link[=sample_index]{sample_index()}}
}
