% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sxpdb.R
\name{show_query}
\alias{show_query}
\title{Displays a query}
\usage{
show_query(query)
}
\arguments{
\item{query}{query object}
}
\value{
integer, number of defined parameters
}
\description{
\code{show_query} displays the content of a query, i.e. all its parameters.
If a parameter is not defined, it shows \code{any} for its value.
For class names, it shows a vector of class names, possibly an empty one.
}
\seealso{
\code{\link[=query_from_value]{query_from_value()}}, \code{\link[=query_from_plan]{query_from_plan()}}, \code{\link[=relax_query]{relax_query()}}, \code{\link[=close_query]{close_query()}}
}
