% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sxpdb.R
\name{size_db}
\alias{size_db}
\title{Number of elements in the database}
\usage{
size_db(db)
}
\arguments{
\item{db}{database, sxpdb object}
}
\value{
integer, number of values in the database, or \code{NULL} in case of error
}
\description{
\code{size_db} returns the number of elements in the database. All elements are unique.
}
\seealso{
\code{\link[=nb_values_db]{nb_values_db()}}
}
