% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sxpdb.R
\name{string_sexp_type}
\alias{string_sexp_type}
\title{Get a textual repreentation of an R type}
\usage{
string_sexp_type(int_type)
}
\arguments{
\item{int_type}{integer representation of a type}
}
\value{
character vector
}
\description{
\code{string_sexp_type} shows the character representation of an R type encoded
as an integer.
}
