% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sxpdb.R
\name{values_from_calls}
\alias{values_from_calls}
\title{Fetches all the calls corresponding to one origin}
\usage{
values_from_calls(db, pkg_name, fun_name)
}
\arguments{
\item{db}{database, sxpdb object}

\item{pkg_name}{character vector of the name of the package}

\item{fun_name}{character vector of the name of the function}
}
\value{
data frame with the following columns:
\itemize{
\item \code{call_id}: unique id of the call
\item \code{value_id}: unique id of the value
\item \code{param}: parameter name
}
}
\description{
\code{values_from_calls}  fetches the values that share the given origin, i.e.
that are one of the arguments or the return value of \code{pkg_name::fun_name}, and
identify from which calls they come from.
}
\seealso{
\code{\link[=values_from_origin]{values_from_origin()}}, \code{\link[=view_origins_db]{view_origins_db()}}, \code{\link[=view_call_ids]{view_call_ids()}}
}
