% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sxpdb.R
\name{values_from_origin}
\alias{values_from_origin}
\title{Fetches all the values corresponding to one origin}
\usage{
values_from_origin(db, pkg_name, fun_name)
}
\arguments{
\item{db}{database, sxpdb object}

\item{pkg_name}{character vector of the name of the package}

\item{fun_name}{character vector of the name of the function}
}
\value{
data frame with two columns:
\itemize{
\item \code{id}: ids of the values
\item \code{param}: parameter names, concatenated and separated with \verb{;}
}
}
\description{
\code{values_from_origins} fetches the values that share the given origin, i.e.
that are one of the arguments or the return value of \code{pkg_name::fun_name}.
}
\seealso{
\code{\link[=view_origins_db]{view_origins_db()}}, \code{\link[=values_from_calls]{values_from_calls()}}
}
