% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sxpdb.R
\name{view_call_ids}
\alias{view_call_ids}
\title{Fetches call ids from the database.}
\usage{
view_call_ids(db, query = NULL)
}
\arguments{
\item{db}{database, sxpdb object}

\item{query}{not taken into account currently}
}
\value{
a data frame with column \code{call_id} where each cell is a list of integers, the call ids, and each row
corresponds to a value. The values are ordered in the same way as their indexes in the database.
}
\description{
\code{view_call_ids} fetches call ids from the database, according to a given query.
}
\seealso{
\code{\link[=map_db]{map_db()}} \code{\link[=get_meta_idx]{get_meta_idx()}} \code{\link[=view_db]{view_db()}} \code{\link[=map_db]{map_db()}} \code{\link[=filter_index_db]{filter_index_db()}} \code{\link[=view_db_names]{view_db_names()}}
}
