% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sxpdb.R
\name{view_db}
\alias{view_db}
\title{Fetch values from the database.}
\usage{
view_db(db, query = NULL)
}
\arguments{
\item{db}{database, sxpdb object}

\item{query}{query object, typically built from \code{\link[=query_from_plan]{query_from_plan()}} or \code{\link[=query_from_value]{query_from_value()}}.}
}
\value{
list of values matching the query
}
\description{
\code{view_db} fetches values from the database, given a query. It will materialize the R values,
and might quickly feel up memory. Rather use metadata (with \code{\link[=view_meta_db]{view_meta_db()}}) if you don't need
to look at the valeus itself. If you need to extract data from the values or transform them, rather use
\code{\link[=map_db]{map_db()}}, which will only load one value at a time.
}
\seealso{
\code{\link[=map_db]{map_db()}} \code{\link[=view_meta_db]{view_meta_db()}} \code{\link[=filter_index_db]{filter_index_db()}} \code{\link[=get_value_idx]{get_value_idx()}}
}
