% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sxpdb.R
\name{view_db_names}
\alias{view_db_names}
\title{Fetches the origin databases}
\usage{
view_db_names(db, query = NULL)
}
\arguments{
\item{db}{database, sxpdb object}

\item{query}{not taken into account currently}
}
\value{
a data frame with column \code{dbname} with the database names, in the same order as the values
ordered by indexes in the database.
}
\description{
\code{view_db_names} fetches the origin databases of each values, i.e. in which database they were written
first, if they result from merging in the current database.
}
\seealso{
\code{\link[=map_db]{map_db()}} \code{\link[=get_meta_idx]{get_meta_idx()}} \code{\link[=view_db]{view_db()}} \code{\link[=map_db]{map_db()}} \code{\link[=filter_index_db]{filter_index_db()}} \code{\link[=view_call_ids]{view_call_ids()}}
}
