% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sxpdb.R
\name{view_meta_db}
\alias{view_meta_db}
\title{Fetches metadata from the database.}
\usage{
view_meta_db(db, query = NULL)
}
\arguments{
\item{db}{database, sxpdb object}

\item{query}{query object, typically built from \code{\link[=query_from_plan]{query_from_plan()}} or \code{\link[=query_from_value]{query_from_value()}}.}
}
\value{
data frame of the metadata for all the values, in the order of their indexes in the database
}
\description{
\code{view_meta_db} fetches metadata from the database, given a query.
They include:
* \emph{runtime} metadata: number of calls the value has been seen in, number of times the value has been seen during merges
* \emph{static} metadata: type, size in bytes, length (for vector values), number of attributes, number of dimensionsm number of rows (for data frames, matrixes)
* \emph{debug} metadata: only if the database was created with sxpdb in debug mode, includes how many times \code{MAYBE_SHARED} has been true on the value,
and how many times we were able to use the SEXP address optimization
}
\seealso{
\code{\link[=map_db]{map_db()}} \code{\link[=get_meta_idx]{get_meta_idx()}} \code{\link[=view_db]{view_db()}} \code{\link[=map_db]{map_db()}} \code{\link[=filter_index_db]{filter_index_db()}}
}
