% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sxpdb.R
\name{view_origins_db}
\alias{view_origins_db}
\title{Fetches origins from the database}
\usage{
view_origins_db(db, query = NULL)
}
\arguments{
\item{db}{database, sxpdb object}

\item{query}{query object, typically built from \code{\link[=query_from_plan]{query_from_plan()}} or \code{\link[=query_from_value]{query_from_value()}}.}
}
\value{
data frame with columns \code{id}, \code{pkg}, \code{fun} and \code{param} for the index in the database, and package
function and parameter names.
}
\description{
\code{view_origins_db} fetches origins from the database, according to a given query.
}
\seealso{
\code{\link[=get_origins]{get_origins()}}, \code{\link[=get_origins_idx]{get_origins_idx()}}, \code{\link[=view_db]{view_db()}}, \code{\link[=map_db]{map_db()}}, \code{\link[=view_meta_db]{view_meta_db()}}
}
