% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sxpdb.R
\name{write_mode}
\alias{write_mode}
\title{Checks if the database is in write mode}
\usage{
write_mode(db)
}
\arguments{
\item{db}{database, sxpdb object}
}
\value{
boolean, \code{TRUE} if it is in write mode
}
\description{
The database can be open in read mode, write mode or merge mode. \code{write_mode} checks if it is open
in write mode.
}
\seealso{
\code{\link[=open_db]{open_db()}}
}
